Enable verbose output for jscpd
`jscpd` exits with exit code 0 even if issues were detected, which is in conflict with pre-commit's expectations.
I opened a pull request with a proposed fix - https://github.com/kucherenko/jscpd/pull/463.
Until that is merged, we would have to use the verbose output for `jscpd`. This won't stop the commit in case of errors, but would at least show visual feedback.